requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
websockets==12.0
